backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
supabase>=2.0.0
pydantic>=2.0.0
google-genai>=0.3.0
yfinance>=0.2.0
numpy>=1.24.0
python-json-logger>=2.0.0
pytz>=2024.1

